Added this File Via Upload
This is a Problem in which we have to Input a Number "N", which is the Number of Test cases and 0<N<=10, We have to Insert this many Strings, and the Length of the Strings can be anywhere between 1 to 10000, We have to print the Even place and Odd Place Characters is a way: EvenCharacter  OddCharacters. this must be done for all the Strings (i.e. for N times in separate lines).